// plugins/org.eclipse.osee.web/src/app/ple/plconfig/lib/types/pl-config-cfggroups.ts
/*********************************************************************
 * Copyright (c) 2021 Boeing
 *
 * This program and the accompanying materials are made
 * available under the terms of the Eclipse Public License 2.0
 * which is available at https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *     Boeing - initial API and implementation
 **********************************************************************/
import { view } from './pl-config-applicui-branch-mapping';

export interface addCfgGroup {
	title: string;
	description: string;
}

export interface CfgGroupDialog {
	configGroup: {
		name: string;
		description: string;
		id: string;
		views: view[];
		configurations: string[];
	};
	editable: boolean;
}
export interface ConfigurationGroupDefinition {
	id?: string;
	name: string;
	description: string;
	configurations?: string[];
}